CRUD working with history of statuses data

+++ src/pages/api/patients.ts
@@ -9,7 +9,7 @@ const supabase = createClient(supabaseUrl!, supabaseKey!);
 
 type ResponseData = {
   message: string;
-  patients?: Patient[];
+  patients: Patient[];
 };
 
 export default async function handler(
@@ -17,13 +17,18 @@ export default async function handler(
   res: NextApiResponse<ResponseData>
 ) {
   if (req.method === 'GET') {
-    // read from database
+    // read all patients and status updates from database
+    const { data: items, error } = await supabase
+      .from('patients')
+      .select('*, status_update(*)');
 
-    const { data: items, error } = await supabase.from('patients').select('*');
     if (error) {
-      res.status(500).json({ message: 'Error fetching items', patients: [] });
+      res
+        .status(500)
+        .json({ message: `Error fetching items: ${error}`, patients: [] });
       return;
     }
+
     const patients = items.map((item: any) => ({
       id: item.id,
       firstName: item.first_name,
@@ -31,51 +36,102 @@ export default async function handler(
       dateOfBirth: item.date_of_birth,
       status: item.status,
       address: item.address,
+      providerId: item.provider_id,
+      statusUpdates: item.status_update,
     }));
+
     res
       .status(200)
       .json({ message: 'Patients fetched successfully', patients });
   } else if (req.method === 'POST') {
-    // write to database
+    // add patient and status update to database
     const patient =
       typeof req.body === 'string' ? JSON.parse(req.body) : req.body;
 
-    const { data, error } = await supabase
+    const { data: patientData, error: patientError } = await supabase
       .from('patients')
       .insert([
         {
-          id: patient.id,
           first_name: patient.firstName,
           middle_name: patient.middleName,
           last_name: patient.lastName,
           date_of_birth: patient.dateOfBirth,
           status: patient.status,
           address: patient.address,
+          provider_id: patient.providerId,
         },
       ])
       .select();
 
-    if (error) {
+    if (patientError || !patientData || patientData.length === 0) {
       res.status(500).json({ message: 'Error adding patient', patients: [] });
       return;
     }
+    const newPatient: Patient = {
+      id: patientData[0].id,
+      firstName: patientData[0].first_name,
+      lastName: patientData[0].last_name,
+      dateOfBirth: patientData[0].date_of_birth,
+      status: patientData[0].status,
+      address: patientData[0].address,
+      providerId: patientData[0].provider_id,
+      statusUpdates: [],
+    };
+    const { data: statusUpdateData, error: statusUpdateError } = await supabase
+      .from('status_update')
+      .insert([{ patient_id: newPatient.id, status: newPatient.status }])
+      .select();
+
+    if (
+      statusUpdateError ||
+      !statusUpdateData ||
+      statusUpdateData.length === 0
+    ) {
+      res.status(500).json({
+        message: 'Error adding status update',
+        patients: [],
+      });
+      return;
+    }
+    const newStatusUpdate = {
+      patientId: statusUpdateData[0].patient_id,
+      status: statusUpdateData[0].status,
+      createdAt: statusUpdateData[0].created_at,
+    };
+
     res.status(200).json({
       message: 'Patient added successfully',
-      patients: data.map((item: any) => ({
-        id: item.id,
-        firstName: item.first_name,
-        lastName: item.last_name,
-        dateOfBirth: item.date_of_birth,
-        status: item.status,
-        address: item.address,
-      })),
+      patients: [{ ...newPatient, statusUpdates: [newStatusUpdate] }],
     });
   } else if (req.method === 'PUT') {
     // update in database
     const patient =
       typeof req.body === 'string' ? JSON.parse(req.body) : req.body;
 
-    const { error } = await supabase
+    if (patient.isStatusUpdate) {
+      const { data: statusUpdateData, error: statusUpdateError } =
+        await supabase
+          .from('status_update')
+          .insert([
+            {
+              patient_id: patient.id,
+              status: patient.status,
+            },
+          ])
+          .select();
+      if (
+        statusUpdateError ||
+        !statusUpdateData ||
+        statusUpdateData.length === 0
+      ) {
+        res
+          .status(500)
+          .json({ message: 'Error updating status update', patients: [] });
+        return;
+      }
+    }
+
+    const { data: patientData, error: patientError } = await supabase
       .from('patients')
       .update({
         first_name: patient.firstName,
@@ -85,30 +141,57 @@ export default async function handler(
         status: patient.status,
         address: patient.address,
       })
-      .eq('id', patient.id);
+      .eq('id', patient.id)
+      .select('*, status_update(*)');
 
-    if (error) {
+    if (patientError) {
       res.status(500).json({ message: 'Error updating patient', patients: [] });
       return;
     }
-    res.status(200).json({ message: 'Patient updated successfully' });
+    const updatedPatient: Patient = {
+      id: patientData[0].id,
+      firstName: patientData[0].first_name,
+      lastName: patientData[0].last_name,
+      dateOfBirth: patientData[0].date_of_birth,
+      status: patientData[0].status,
+      address: patientData[0].address,
+      providerId: patientData[0].provider_id,
+      statusUpdates: patientData[0].status_update,
+    };
+    res.status(200).json({
+      message: 'Patient updated successfully',
+      patients: [updatedPatient],
+    });
   } else if (req.method === 'DELETE') {
     // delete from database
     const patient =
       typeof req.body === 'string' ? JSON.parse(req.body) : req.body;
 
-    const { error } = await supabase
+    const { error: statusUpdateError } = await supabase
+      .from('status_update')
+      .delete()
+      .eq('patient_id', patient.id);
+
+    if (statusUpdateError) {
+      res
+        .status(500)
+        .json({ message: 'Error deleting status update', patients: [] });
+      return;
+    }
+
+    const { error: patientError } = await supabase
       .from('patients')
       .delete()
       .eq('id', patient.id);
 
-    if (error) {
+    if (patientError) {
       res.status(500).json({ message: 'Error deleting patient', patients: [] });
       return;
     }
 
     res.status(200).json({
       message: 'Patient deleted successfully',
+      patients: [],
     });
   }
 }
